Move decrypt_file to travis, since it should be run on Travis
Also, fix the Sphinx references in the encrypt_file and decrypt_file
docstrings to each other.
