+++ doctr/__init__.py
@@ -1,11 +1,11 @@
-from .local import encrypt_variable, encrypt_file, decrypt_file, generate_GitHub_token
-from .travis import (get_token, run, setup_GitHub_push, gh_pages_exists,
-    create_gh_pages, commit_docs, push_docs, get_repo)
+from .local import encrypt_variable, encrypt_file, generate_GitHub_token
+from .travis import (decrypt_file, get_token, run, setup_GitHub_push,
+    gh_pages_exists, create_gh_pages, commit_docs, push_docs, get_repo)
 
 __all__ = [
-    'encrypt_variable', 'encrypt_file', 'decrypt_file', 'generate_GitHub_token',
+    'encrypt_variable', 'encrypt_file', 'generate_GitHub_token',
 
-    'get_token', 'run', 'setup_GitHub_push', 'gh_pages_exists',
+    'decrypt_file', 'get_token', 'run', 'setup_GitHub_push', 'gh_pages_exists',
     'create_gh_pages', 'commit_docs', 'push_docs', 'get_repo',
 ]
