+++ doctr/local.py
@@ -59,7 +59,7 @@ def encrypt_file(file):
 
     Returns the secret key used for the encryption.
 
-    Decrypt the file with :func:`decrypt_file`.
+    Decrypt the file with :func:`doctr.travis.decrypt_file`.
 
     """
     key = Fernet.generate_key()
@@ -72,32 +72,6 @@ def encrypt_file(file):
         f.write(encrypted_file)
 
     return key
-
-def decrypt_file(file, key):
-    """
-    Decrypts the file ``file``.
-
-    The encrypted file is assumed to end with the ``.enc`` extension. The
-    decrypted file is saved to the same location without the ``.enc``
-    extension.
-
-    The permissions on the decrypted file are automatically set to 0o600.
-
-    See also :func:`encrypt_file`.
-
-    """
-    if not file.endswith('.enc'):
-        raise ValueError("%s does not end with .enc" % file)
-
-    fer = Fernet(key)
-
-    with open(file, 'rb') as f:
-        decrypted_file = fer.decrypt(f.read())
-
-    with open(file[:4], 'wb') as f:
-        f.write(decrypted_file)
-
-    os.chmod(file[:4], 0o600)
 
 class AuthenticationFailed(Exception):
     pass
